+++ docs/HISTORY.md
@@ -3,7 +3,8 @@
 ## 4.0.0+slc.2 (unreleased)
 
 
-- Nothing changed yet.
+- Deprecate the old namespace.
+  [ale-rt]
 
 
 ## 4.0.0+slc.1 (2026-02-10)